# repository/models.go
// Code generated by sqlc. DO NOT EDIT.

package repository

import (
	"time"

	"github.com/google/uuid"
)

type Account struct {
	AccountID   uuid.UUID `json:"account_id"`
	Username    string    `json:"username"`
	Password    string    `json:"password"`
	Email       string    `json:"email"`
	AccountType string    `json:"account_type"`
	CreatedAt   time.Time `json:"created_at"`
	UpdatedAt   time.Time `json:"updated_at"`
}

type Order struct {
	OrderID     int64     `json:"order_id"`
	OrderPrice  string    `json:"order_price"`
	Idr         string    `json:"idr"`
	Btc         string    `json:"btc"`
	OrderType   string    `json:"order_type"`
	OrderCrypto string    `json:"order_crypto"`
	OrderStatus string    `json:"order_status"`
	CreatedAt   time.Time `json:"created_at"`
	UpdatedAt   time.Time `json:"updated_at"`
}